Refactored Interaction System
