+++ Assets/Scripts/CharacterHandler.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using GameConstants;
-using Movement;
 using UnityEngine;
 using UnityEngine.Events;
 using UnityEngine.InputSystem;
@@ -35,9 +34,14 @@ public class CharacterHandler : MonoBehaviour
     private PlayerInput player1Input;
     private PlayerInput player2Input;
 
+    public InputMode GhostInputMode { get; set; }
+    public InputMode HumanInputMode { get; set; }
+
     private void Start()
     {
         movementInputEnabled = true;
+        GhostInputMode = InputMode.Free;
+        HumanInputMode = InputMode.Free;
     }
 
     private void Update()
