+++ Assets/Scripts/InputMode.cs
@@ -1,0 +1,5 @@
+﻿public enum InputMode
+{
+    Free,
+    InQTE
+}
