+++ Assets/Scripts/Interaction.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9eeb45964b524e6d84449550c5a68479
+timeCreated: 1675853131
