+++ Assets/Scripts/Interaction/Interactable.cs.meta
